al2: Fix benchmark 5.4.4 (umask) to pass Amazon Inspector.
Even though inspection of `/etc/profile` and `/etc/bashrc` allow it to
be obvious that control flow passes over the umask lines added to the
end of the file, both the CIS Benchmark scoring instructions as well as
Amazon Inspector require that every `umask` present in the relevant
files are updated.

By adding this change, Amazon Inspector will successfully pass the CIS
Benchmark controls related to having correctly restricted the umask.