al2: Correctly override motd banner (1.7.1.1).
While the `update-motd` cron was removed, there is a systemd unit that
executes on VM start which re-computes the motd file and replaces the
contents. Instead of disabling this systemd unit, we instead replace the
contents of the script which is used for generating the motd.

By adding this change, Amazon Inspector will successfully pass the CIS
Benchmark controls related to having correctly set `/etc/motd`.