mounting: Add nodev and nosuid mount options to var partitions.
According to CIS Benchmark control 1.1.13, `/home` should be mounted
`nodev`. This updates how the `/etc/fstab` file is constructed so this
passes Amazon Inspector.

In addition to setting `nodev` on `/home`, it seems reasonable to also
apply this setting across the `/var`, `/var/log`, and `/var/log/audit`
mounts as it is unlikely `/var` would need devices, and increasingly
less likely for `/var/log` and `/var/log/audit`.

Lastly, `nosuid` was applied to the `/home`, `/var/log`, and
`/var/log/audit` mounts. It seems fairly reasonable that the two log
mounts won't need anything suid related, and fairly unlikely that
`/home` would need it either. It is likely this could also apply to
`/var`, but it is difficult to be certain.